+++ src/api/handlers/index.js
@@ -1,3 +1,3 @@
 export { default as ingress } from './ingress'
-export { default as twiml } from './twiml'
-export { default as call } from './call'
+export { default as twiml } from './twilio/twiml'
+export { default as call } from './twilio/call'
